Create backup.sh
This file is for using rsync to back up differences on a schedule (with cron perhaps), and with automatic login to the root account of a working machine. When the backup is complete, it will replace the "current" with the latest in order to provide a new target to compare against.  To start you must create a blank "current" directory for it to compare to.